refactor: split Android native perf collector
